+++ app/Main.hs
@@ -6,6 +6,7 @@ import Data.Either (lefts, rights)
 import Data.List.NonEmpty (toList)
 import Evaluation (Value, evalExpr, printValue)
 import Expression (Expression, expression, prettyPrintExpr)
+import Grammar (Program, program)
 import Parser (runParser)
 import Scanner (SyntaxError, prettyPrintErr, scanTokens)
 import System.Environment (getArgs)
@@ -21,8 +22,8 @@ main = do
     ["--chap04_scanning", script] -> readFile' script >>= handleChap04Out . runChapter04
     ["--chap06_parsing", script] -> readFile' script >>= handleChap06Out . (runChapter04 >=> runChapter06)
     ["--chap07_evaluating", script] -> readFile' script >>= handleChap07Out . (runChapter04 >=> runChapter06 >=> runChapter07)
-    ["--chap08_statements", script] -> readFile' script >>= handleChap07Out . currentImpl -- TODO same as chapter 07 for now
-    [script] -> readFile' script >>= handleChap07Out . currentImpl
+    ["--chap08_statements", script] -> readFile' script >>= handleChap08Out . currentImpl -- TODO same as chapter 08 for now
+    [script] -> readFile' script >>= handleChap08Out . currentImpl
     _ -> do
       putStrLn "Usage: hox [[--<CHAP>] script]"
       exitWith (ExitFailure 64)
@@ -30,9 +31,10 @@ main = do
     handleChap04Out = either handleErr (mapM_ (putStrLn . prettyPrintToken))
     handleChap06Out = either handleErr (putStrLn . prettyPrintExpr)
     handleChap07Out = either handleErr (putStrLn . printValue)
+    handleChap08Out = either handleErr print -- TODO improve
 
-currentImpl :: String -> Either InterpreterError Value
-currentImpl = runChapter04 >=> runChapter06 >=> runChapter07
+currentImpl :: String -> Either InterpreterError Program
+currentImpl = runChapter04 >=> runChapter08
 
 -- Chapter 04 operations
 runChapter04 :: String -> Either InterpreterError [Token]
@@ -50,6 +52,11 @@ runChapter06 = bimap Parse fst . runParser expression
 -- Chapter 07 operations
 runChapter07 :: Expression -> Either InterpreterError Value
 runChapter07 = first Eval . evalExpr
+
+-- Chapter08 operations
+-- The previous chapters where "but a hack". Now we have the real deal
+runChapter08 :: [Token] -> Either InterpreterError Program
+runChapter08 = bimap Parse fst . runParser program
 
 -- Error handling
 data InterpreterError = Syntax [SyntaxError] | Parse String | Eval String